Сделал CRUD для всех таблиц. Исправил несколько ошибок.

+++ PensionAccumulationCalculator/Entities/Individual_pencion_coefficient_accumulation.cs
@@ -4,7 +4,7 @@
         public int User_id { get; set; }
         public int Record_id { get; set; } 
         public required string Record_name { get; set; }
-        public float Individual_pencion_coefficient { get; set; } 
+        public decimal Individual_pencion_coefficient { get; set; } 
         public int Year { get; set; }
     }
 }
