+++ PensionAccumulationCalculator/Entities/Insurance_record.cs
@@ -2,7 +2,7 @@
     public class Insurance_record {
         public int Insurance_exp_id { get; set; }
         public int User_id { get; set; }
-        public float Individual_pension_coefficient { get; set; }
+        public decimal Individual_pension_coefficient { get; set; }
         public int Year {  get; set; }
     }
 }
